dom/canvas/test/test_imagebitmap_transfer.html

<!DOCTYPE HTML>
<title>Test ImageBitmap : Transfer</title>
<meta charset="utf-8">
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<script src="imagebitmap_structuredclone_utils.js"></script>
<script type="text/javascript">

var worker = new Worker("imagebitmap_structuredclone.js");
worker.onmessage = function(event) {

  if (event.data.type == "status") {
    ok(event.data.status, event.data.msg);
  } else if (event.data.type == "finish") {
    SimpleTest.finish();
  } else if (event.data.type == "bitmap1") {
    compareImageBitmapWithImageElement(event.data.bitmap, gImage1);
  } else if (event.data.type == "bitmap2") {
    compareImageBitmapWithImageElement(event.data.bitmap, gImage2);
  } else if (event.data.type == "bitmap3") {
    compareImageBitmapWithImageData(event.data.bitmap, gImageData, "Check preserving alpha");
  }
}

function runTests() {
  ok(worker, "Worker created successfully.");

  prepareImageBitmaps().then(function(){
    worker.postMessage({"bitmap1":gImageBitmap1, "bitmap2":gImageBitmap2, "bitmap3":gImageBitmap3},
                       [gImageBitmap1, gImageBitmap2, gImageBitmap3]);

    ok(gImageBitmap1.width == 0 && gImageBitmap1.height == 0,
       "After transfer, ImageBitmap become neutered");
    ok(gImageBitmap2.width == 0 && gImageBitmap2.height == 0,
       "After transfer, ImageBitmap become neutered");
    ok(gImageBitmap3.width == 0 && gImageBitmap3.height == 0,
       "After transfer, ImageBitmap become neutered");
  });
}

SimpleTest.waitForExplicitFinish();
addLoadEvent(runTests);

</script>
</body>
